Fix a crash when the mapping process retrieved an error and one tries to remap it manually.
Make script update a tv show, that has season numbering in Kodi, with proper episode number, which are
not in season form in MAL. This change relies on offline information, so one needs to have the complete
former season for this to work.
Change mal-api.ramblingahoge.net to api.atarashiiapp.com, because first one didn't update MAL anymore.